moved TicTacToeGame to header, has issue with moving definitions to source file
