+++ test.cpp
@@ -1,8 +1,9 @@
 
-#define DOCTEST_CONFIG_IMPLEMENT
-#include "doctest.h"
 
 #include "TicTacToeGame.h"
+
+#define DOCTEST_CONFIG_IMPLEMENT
+#include "doctest.h"
 
 int main(int argc, char** argv)
 {
